+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -1,6 +1,10 @@
-## 0.6.21+1
+## 0.6.21+2
 
 * Bumps com.android.tools.build:gradle from 8.12.1 to 8.13.0 and kotlin-gradle-plugin from 2.2.10 to 2.2.20.
+
+## 0.6.21+1
+
+* Updates to stable version of `androidx.camera` 1.5.
 
 ## 0.6.21