+++ packages/google_sign_in/google_sign_in_android/CHANGELOG.md
@@ -1,6 +1,10 @@
-## 7.0.6
+## 7.1.1
 
 * Bumps com.android.tools.build:gradle from 8.12.1 to 8.13.0 and kotlin-gradle-plugin from 2.2.10 to 2.2.20.
+
+## 7.1.0
+
+* Adds support for the `clearAuthorizationToken` method.
 
 ## 7.0.5